Let's see, a lot has changed. Should have split these changes in multiple commits =(
Changed how command arguments are parsed into the derired types.
Current implementation now allows for containers that implement
the push_back() method. Values for these containers are represented
with json arrays. The jsoncons header-only library was used to do
all the json stuff.

Reimplemented the shell so it no longer uses cpp-readline, and just
uses readline directly. It's better now because some functionality
of cpp-readline was redundant with shpp's.

Parameter types can now be printed and shown, and the shell can
print a command's signature. The help command now shows a list of
command signatures instead of just command names.